se mueve y busca comida pero maso maso

+++ TP3/Geometry.h
@@ -4,7 +4,7 @@
 #define PI 3.14159265
 
 #include "Point.h"
-int distanceB2Points(Point& p1, Point& p2);
+double distanceB2Points(Point& p1, Point& p2);
 double getAngleBetweenPoints(Point& p1, Point& p2);
 Point translatePoint(Point p, double distance, double angle);
